requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
django==4.2.23
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
exceptiongroup==1.3.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
pygments==2.19.2
PyJWT==2.9.0
pytest==8.4.1
requests==2.32.4
sqlparse==0.5.3
tomli==2.2.1
typing-extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
